+++ r-package/R/travel_time_matrix.R
@@ -22,7 +22,7 @@
 #'   specified time window are shorter than 15 minutes. Defaults to 50,
 #'   returning the median travel time. If a vector with length bigger than 1 is
 #'   passed, the output contains an additional column for each percentile
-#'   specifying the percentile travel time estimate. each estimate. Due to
+#'   specifying the percentile travel time estimate. Due to
 #'   upstream restrictions, only 5 percentiles can be specified at a time. For
 #'   more details, please see R5 documentation at
 #'   <https://docs.conveyal.com/analysis/methodology#accounting-for-variability>.
@@ -166,16 +166,21 @@ travel_time_matrix <- function(r5r_network,
   departure <- assign_departure(departure_datetime)
 
   # check availability of transit services on the selected date
-  if (mode_list$transit_mode %like% 'TRANSIT|TRAM|SUBWAY|RAIL|BUS|CABLE_CAR|GONDOLA|FUNICULAR') {
+  if (mode_list$transit_mode %like% 'TRANSIT|TRAM|SUBWAY|RAIL|BUS|FERRY|CABLE_CAR|GONDOLA|FUNICULAR') {
     check_transit_availability_on_date(r5r_network, departure_date = departure$date)
   }
 
   checkmate::assert_class(r5r_network, "r5r_network")
   r5r_network <- r5r_network@jcore
 
-  # in direct modes reverse origin/destination to take advantage of R5's One to Many algorithm
+  # in direct modes reverse origin/destination to take advantage of R5's One to Many algorithm.
+  # skipped when output_dir is set: Java writes the CSVs with the swapped from_id/to_id and names
+  # the files after the swapped origins, and the swap is only undone in the in-memory result
   data_path <- r5r_network$getDataPath()
-  res <- reverse_if_direct_mode(origins, destinations, mode_list, data_path)
+  res <- NULL
+  if (is.null(output_dir)) {
+    res <- reverse_if_direct_mode(origins, destinations, mode_list, data_path)
+  }
   if (!is.null(res)) {
     origins <- res$origins
     destinations <- res$destinations
@@ -218,7 +223,7 @@ travel_time_matrix <- function(r5r_network,
   set_verbose(r5r_network, verbose)
   set_progress(r5r_network, progress)
   set_fare_structure(r5r_network, fare_structure)
-  set_max_fare(r5r_network, max_fare)
+  set_max_fare(r5r_network, max_fare, fare_structure)
   set_output_dir(r5r_network, output_dir)
   set_expanded_travel_times(r5r_network, FALSE)
   set_breakdown(r5r_network, FALSE)